agent/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.1
